refactor(api): single responsibility api
