+++ src/api/index.js
@@ -1,0 +1,4 @@
+import { getIngredients } from "./rest/ingredients";
+import { getOrder } from "./rest/orders";
+
+export { getIngredients, getOrder };
