+++ src/components/burger-ingredients/burger-ingredients.jsx
@@ -3,9 +3,9 @@ import Modal from "../ui/modal/modal";
 import IngredientDetails from "../ui/ingredient-details/ingredient-details";
 import { ingredientsSlice } from "../../services/slices/ingredients";
 import { useDispatch } from "react-redux";
-import { getIngredients } from "../../utils/api";
 import TabIngredients from "../ui/tab-ingredients/tab-ingredients";
 import { useSelector } from "react-redux";
+import { getIngredients } from "../../api";
 
 const BurgerIngredients = () => {
   const { ingredients, ingredientsFailed, ingredientsRequest } = useSelector(
@@ -31,12 +31,13 @@ const BurgerIngredients = () => {
       {ingredientsRequest ? (
         "Загрузка ингредиентов..."
       ) : ingredientsFailed ? (
-        "Ошибка сервера, зайдите позже"
+        "Ошибка сервера, зайдите позже."
       ) : ingredients.length <= 0 ? (
         "Извините, но у нас закончились продукты, зайдите позже."
       ) : (
         <>
           <TabIngredients openIngredient={openIngredient} />
+
           {openModal && (
             <Modal onClose={() => setOpenModal(false)}>
               <IngredientDetails />
