--- src/utils/utils.js
@@ -1,5 +1,0 @@
-export const checkResponse = (response) => {
-  if (!response.ok)
-    throw new Error(`${response.status} — ${response.statusText}`);
-  return response.json();
-};
